requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
google==3.0.0
idna==2.10
Pillow==8.2.0
pypng==0.0.20
PyQRCode==1.2.1
pyshorteners==1.0.1
pyzbar==0.1.8
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.5
